+++ model/src/main/java/jm/model/ChannelDTO.java
@@ -1,0 +1,27 @@
+package jm.model;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+
+
+@Data
+@NoArgsConstructor
+@Entity
+public class ChannelDTO {
+
+    @Id
+    @Column(name = "id", nullable = false, unique = true, updatable = false)
+    private Long id;
+
+    @Column(name = "name", nullable = false)
+    private String name;
+
+    public ChannelDTO(Channel channel) {
+        this.id = channel.getId();
+        this.name = channel.getName();
+    }
+}
